Fixing a bug when bootstrapping work machine
When installing docker with apt (stretch), docker tries to start using
init script and fails because systemd socket's unit is not started along
with it. It finally succeeds because systemd retries later and this
time, starts dependent units (socket). So if we restart bootstrap, it
finally succeeds...
This hack permits the machine to be bootstrapped alone without any
exterior help.